requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
bcrypt==3.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi-mail==1.4.1
fastapi==0.120.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
mypy==1.18.2
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pytest==8.4.2
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
types-passlib==1.7.7.20250602
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
sqladmin
jinja2==3.1.4
python-multipart
websockets
pytest
